Linux command- pushd-popd comments
Stack Data Structure: The stack is used to store directory paths. It has functions to initialize, check if it is full or empty, push a directory onto the stack, and pop a directory from the stack.pushd:Retrieves the current working directory using getcwd.Pushes this directory onto the stack.Changes to the new directory specified by the user using chdir.popd:Pops the top directory from the stack.Changes to this directory using chdir.Main Function:Validates the input arguments.Calls pushd or popd based on the command provided.Prints the current directory after the operation.